Revert "qemuarma9: Add uImage as KERNEL_ALT_IMAGETYPE"
This reverts commit 1f4c9457250a4eb8fd5748a964029114641ca72e.

The mkimage tool is currently needed on this host to build the
qemuarma9 board. Not all hosts have mkimage installed  so
most qemuarma9 builds break with this error:

  Kernel: arch/arm/boot/zImage is ready
  UIMAGE  arch/arm/boot/uImage
"mkimage" command not found - U-Boot images will not be built
/.../bitbake_build/tmp/work-shared/qemuarma9/kernel-source/arch/arm/boot/Makefile:83: recipe for target 'arch/arm/boot/uImage' failed
make[3]: *** [arch/arm/boot/uImage] Error 1

Temporarily remove uImage from the BSP layer, since it is just
used for booting on Simics.