qemuarma9: qemu.inc in qemuarma9 BSP obscures qemu.inc in oe-core
Issues: LIN9-1231

The structure of qemuarma9 BSP is incorrect, and it causes issues such
as LIN8-2691 where the qemu.inc defined locally obscures the oe-core
file with the same name.

In LIN8-2691 qemuarma9 must inherit the XSERVER packages defined by
qemu-core, but it is prevented from doing so by the local qemu.inc.

Signed-off-by: Alexandru Moise <alexandru.moise@windriver.com>
Signed-off-by: Liang Li <liang.li@windriver.com>